requirements.txt
# Core
torch>=2.3
transformers>=4.45
peft>=0.13
accelerate>=0.34
bitsandbytes>=0.43
datasets>=2.21
sentencepiece>=0.2
tokenizers>=0.20
numpy>=1.26
pyyaml>=6.0
tqdm>=4.66

# Training fast-path (pin: Unsloth API moves fast)
unsloth==2024.12.4
unsloth-zoo==2024.12.5

# Inference (REQUIRED — not optional)
vllm>=0.6.3

# Judge providers
anthropic>=0.39
openai>=1.50

# Plotting (headline figure)
matplotlib>=3.9

# Flash Attention 2 — install separately with:
#   pip install flash-attn --no-build-isolation
# (Requires CUDA toolchain; see README.)
